Interface for ChartCompiler
