+++ src/test/java/chart/csv/CsvChartCompilerTest.java
@@ -1,4 +1,4 @@
-package chart;
+package chart.csv;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
@@ -10,34 +10,40 @@ import java.util.ArrayList;
 import org.joda.time.DateTime;
 import org.junit.Test;
 
-import chart.csv.CsvSimpleChart;
-import chart.csv.CsvSimpleChartEntry;
-import chart.csv.ImmutableCsvSimpleChart;
-import chart.csv.ImmutableCsvSimpleChartEntry;
+import chart.Chart;
+import chart.ChartCompiler;
+import chart.ChartEntry;
+import chart.CsvChart;
+import chart.CsvChartEntry;
+import chart.FileChartReader;
+import chart.ImmutableCsvChart;
+import chart.ImmutableCsvChartEntry;
+import chart.SimpleChart;
+import chart.SimpleChartReader;
 
-public class ChartCompilerTest {
+public class CsvChartCompilerTest {
     private static final CsvSimpleChartEntry SIMPLE_ENTRY = ImmutableCsvSimpleChartEntry.builder()
                                                                                         .title("title")
                                                                                         .artist("artist")
                                                                                         .position(1)
                                                                                         .build();
     private static final CsvChartEntry CHART_ENTRY = ImmutableCsvChartEntry.builder()
-                                                                              .position(1)
-                                                                              .title("title")
-                                                                              .artist("artist")
-                                                                              .weeksOnChart(1)
-                                                                              .build();
+                                                                           .position(1)
+                                                                           .title("title")
+                                                                           .artist("artist")
+                                                                           .weeksOnChart(1)
+                                                                           .build();
     private static final CsvSimpleChartEntry OTHER_ENTRY = ImmutableCsvSimpleChartEntry.builder()
                                                                                        .title("other-title").artist("other-artist").position(1).build();
     private static final CsvSimpleChart OTHER_CHART = ImmutableCsvSimpleChart.builder()
                                                                              .week(2).date(DateTime.now()).addEntries(OTHER_ENTRY).build();
-    public static final DateTime DEFAULT_DATE = new DateTime(2017, 1, 1, 0, 0);
+    private static final DateTime DEFAULT_DATE = new DateTime(2017, 1, 1, 0, 0);
 
     @Test
     public void canCompileChart() throws IOException {
         FileChartReader reader = new FileChartReader("src/test/resources");
         FileChartReader derivedReader = reader; // TODO
-        ChartCompiler compiler = new ChartCompiler(reader, derivedReader);
+        ChartCompiler compiler = new CsvChartCompiler(reader, derivedReader);
         compiler.compileChart(1);
     }
 
@@ -49,7 +55,7 @@ public class ChartCompilerTest {
 
         FileChartReader derivedReader = mock(FileChartReader.class);
         when(derivedReader.findDerivedChart(0)).thenThrow(IllegalArgumentException.class);
-        ChartCompiler compiler = new ChartCompiler(reader, derivedReader);
+        ChartCompiler compiler = new CsvChartCompiler(reader, derivedReader);
         Chart chart = compiler.compileChart(1);
         assertEquals(1, chart.entries().size());
 
@@ -64,17 +70,17 @@ public class ChartCompilerTest {
 
         FileChartReader derivedReader = mock(FileChartReader.class);
         when(derivedReader.findDerivedChart(1)).thenReturn(defaultChart(1));
-        ChartCompiler compiler = new ChartCompiler(reader, derivedReader);
+        ChartCompiler compiler = new CsvChartCompiler(reader, derivedReader);
         Chart chart = compiler.compileChart(2);
         assertEquals(1, chart.entries().size());
 
         ChartEntry expected = ImmutableCsvChartEntry.builder()
-                .artist("artist")
-                .lastPosition(1)
-                .position(1)
-                .title("title")
-                .weeksOnChart(2)
-                .build();
+                                                    .artist("artist")
+                                                    .lastPosition(1)
+                                                    .position(1)
+                                                    .title("title")
+                                                    .weeksOnChart(2)
+                                                    .build();
 
         assertEquals(expected, chart.entries().get(0));
     }
@@ -87,7 +93,7 @@ public class ChartCompilerTest {
         FileChartReader derivedReader = mock(FileChartReader.class);
         when(derivedReader.findDerivedChart(1)).thenReturn(defaultChart(1));
 
-        ChartCompiler compiler = new ChartCompiler(reader, derivedReader);
+        ChartCompiler compiler = new CsvChartCompiler(reader, derivedReader);
         Chart chart = compiler.compileChart(2);
 
         assertEquals(DEFAULT_DATE.plusWeeks(1), chart.date());
@@ -101,7 +107,7 @@ public class ChartCompilerTest {
 
         FileChartReader derivedReader = mock(FileChartReader.class);
         when(derivedReader.findDerivedChart(1)).thenReturn(defaultChart(1));
-        ChartCompiler compiler = new ChartCompiler(reader, derivedReader);
+        ChartCompiler compiler = new CsvChartCompiler(reader, derivedReader);
         Chart chart = compiler.compileChart(2);
 
         assertEquals(1, chart.dropouts().size());
@@ -115,7 +121,7 @@ public class ChartCompilerTest {
 
         FileChartReader derivedReader = mock(FileChartReader.class);
         when(derivedReader.findDerivedChart(1)).thenReturn(defaultChart(1));
-        ChartCompiler compiler = new ChartCompiler(reader, derivedReader, 42);
+        ChartCompiler compiler = new CsvChartCompiler(reader, derivedReader, 42);
         Chart chart = compiler.compileChart(2);
 
         DateTime expected = DEFAULT_DATE.plusDays(42);
@@ -132,10 +138,10 @@ public class ChartCompilerTest {
 
     private CsvChart defaultChart(int week) {
         return ImmutableCsvChart.builder()
-                .date(DEFAULT_DATE)
-                .week(week)
-                .addEntries(CHART_ENTRY)
-                .dropouts(new ArrayList<>())
-                .build();
+                                .date(DEFAULT_DATE)
+                                .week(week)
+                                .addEntries(CHART_ENTRY)
+                                .dropouts(new ArrayList<>())
+                                .build();
     }
 }
